Refactored MazeGenerator to MazeFactory to eliminate instanciating it
