--- MazeFactory_test.py
@@ -1,4 +1,0 @@
-from MazeFactory import MazeFactory
-
-maze1 = MazeFactory.generate(10, 5)
-print(maze1)
